+++ DockerKafka/docker-compose.yml
@@ -15,7 +15,7 @@ services:
       - "9092:9092"
     environment:
       KAFKA_ADVERTISED_HOST_NAME: localhost
-      KAFKA_CREATE_TOPICS: "kvp-topic:1:1"
+      KAFKA_CREATE_TOPICS: "kvp-input:1:1,kvp-output:1:1"
       KAFKA_ZOOKEEPER_CONNECT: zookeeper:2181
     volumes:
       - /var/run/docker.sock:/var/run/docker.sock
