+++ build.gradle
@@ -1,12 +1,7 @@
 plugins {
 	id 'org.springframework.boot' version '2.5.3'
-	id 'io.spring.dependency-management' version '1.0.11.RELEASE'
 	id 'java'
 }
-
-group = 'com.kvp'
-version = '0.0.1-SNAPSHOT'
-sourceCompatibility = '1.8'
 
 configurations {
 	compileOnly {
@@ -14,22 +9,26 @@ configurations {
 	}
 }
 
-repositories {
-	mavenCentral()
-}
+subprojects {
+	version = '0.0.1-SNAPSHOT'
+	sourceCompatibility = '1.8'
 
-dependencies {
-	implementation 'org.springframework.boot:spring-boot-starter-mustache'
-	implementation 'org.springframework.boot:spring-boot-starter-thymeleaf'
-	implementation 'org.springframework.boot:spring-boot-starter-web'
-	implementation 'org.apache.kafka:kafka-streams'
-	implementation 'org.springframework.kafka:spring-kafka'
-	compileOnly 'org.projectlombok:lombok'
-	annotationProcessor 'org.projectlombok:lombok'
-	testImplementation 'org.springframework.boot:spring-boot-starter-test'
-	testImplementation 'org.springframework.kafka:spring-kafka-test'
-}
+	apply plugin: 'java'
+	apply plugin: 'org.springframework.boot'
+	apply plugin: 'io.spring.dependency-management'
 
-test {
-	useJUnitPlatform()
+	repositories {
+		mavenCentral()
+	}
+
+	dependencies {
+		compileOnly 'org.projectlombok:lombok'
+		annotationProcessor 'org.projectlombok:lombok'
+		implementation 'org.apache.kafka:kafka-streams'
+		implementation 'org.springframework.kafka:spring-kafka'
+	}
+
+	test {
+		useJUnitPlatform()
+	}
 }
