+++ kvp-domain/build.gradle
@@ -1,0 +1,6 @@
+group = 'com.kvp.kafka-domain'
+
+dependencies {
+    testImplementation 'org.junit.jupiter:junit-jupiter-api:5.7.0'
+    testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.7.0'
+}
