+++ kvp-domain/src/main/java/com.kvp.domain/Introduce.java
@@ -1,0 +1,17 @@
+package com.kvp.domain;
+
+import lombok.*;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@ToString
+public class Introduce {
+    private String name;
+    private Long age;
+
+    public void addAge() {
+        age += 10;
+    }
+}
