+++ kvp-kafka/src/main/java/com/kvp/kafka/consumer/ConsumerConfiguration.java
@@ -1,0 +1,38 @@
+package com.kvp.kafka.consumer;
+
+import com.kvp.domain.Introduce;
+import org.apache.kafka.clients.consumer.ConsumerConfig;
+import org.apache.kafka.common.serialization.StringDeserializer;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.kafka.config.ConcurrentKafkaListenerContainerFactory;
+import org.springframework.kafka.core.ConsumerFactory;
+import org.springframework.kafka.core.DefaultKafkaConsumerFactory;
+import org.springframework.kafka.support.serializer.JsonDeserializer;
+
+import java.util.HashMap;
+import java.util.Map;
+
+@Configuration
+public class ConsumerConfiguration {
+    @Bean
+    public ConsumerFactory<String, Introduce> introduceConsumerConfigs() {
+
+        Map<String, Object> configs = new HashMap<>();
+        configs.put(ConsumerConfig.BOOTSTRAP_SERVERS_CONFIG, "localhost:9092");
+        configs.put(ConsumerConfig.GROUP_ID_CONFIG, "kvp");
+        configs.put(ConsumerConfig.AUTO_OFFSET_RESET_CONFIG, "earliest");
+
+        return new DefaultKafkaConsumerFactory<>(
+                configs,
+                new StringDeserializer(),
+                new JsonDeserializer<>(Introduce.class));
+    }
+
+    @Bean
+    public ConcurrentKafkaListenerContainerFactory<String, Introduce> introduceListener() {
+        ConcurrentKafkaListenerContainerFactory<String, Introduce> factory = new ConcurrentKafkaListenerContainerFactory<>();
+        factory.setConsumerFactory(introduceConsumerConfigs());
+        return factory;
+    }
+}
