+++ kvp-kafka/src/main/java/com/kvp/kafka/consumer/KvpTestConsumer.java
@@ -1,6 +1,6 @@
 package com.kvp.kafka.consumer;
 
-import com.kvp.kafka.domain.Introduce;
+import com.kvp.domain.Introduce;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.kafka.annotation.KafkaListener;
 import org.springframework.stereotype.Component;
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Component;
 @Component
 public class KvpTestConsumer {
 
-    @KafkaListener(topics = "kvp-test", groupId = "kvp", containerFactory = "introduceListener")
+    @KafkaListener(topics = "kvp-output", groupId = "kvp", containerFactory = "introduceListener")
     public void consume(Introduce message) {
         log.info("consume message : {}", message);
     }
