+++ settings.gradle
@@ -1,1 +1,5 @@
 rootProject.name = 'kafka'
+include 'kvp-streams'
+include 'kvp-kafka'
+include 'kvp-domain'
+
