--- src/main/java/com/kvp/kafka/domain/Introduce.java
@@ -1,13 +1,0 @@
-package com.kvp.kafka.domain;
-
-import lombok.*;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@ToString
-public class Introduce {
-    private String name;
-    private Long age;
-}
