--- src/main/java/com/kvp/kafka/producer/KafkaProducerConfiguration.java
@@ -1,37 +1,0 @@
-package com.kvp.kafka.producer;
-
-import com.kvp.kafka.domain.Introduce;
-import org.apache.kafka.clients.producer.ProducerConfig;
-import org.apache.kafka.common.serialization.StringSerializer;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.kafka.core.DefaultKafkaProducerFactory;
-import org.springframework.kafka.core.KafkaTemplate;
-import org.springframework.kafka.core.ProducerFactory;
-import org.springframework.kafka.support.serializer.JsonSerializer;
-
-import java.util.HashMap;
-import java.util.Map;
-
-//https://docs.spring.io/spring-kafka/docs/current/reference/html/#kafka-template
-@Configuration
-public class KafkaProducerConfiguration {
-    @Bean
-    public ProducerFactory<String, Introduce> producerFactory() {
-        return new DefaultKafkaProducerFactory<>(producerConfigs());
-    }
-
-    @Bean
-    public Map<String, Object> producerConfigs() {
-        Map<String, Object> props = new HashMap<>();
-        props.put(ProducerConfig.BOOTSTRAP_SERVERS_CONFIG, "localhost:9092");
-        props.put(ProducerConfig.KEY_SERIALIZER_CLASS_CONFIG, StringSerializer.class);
-        props.put(ProducerConfig.VALUE_SERIALIZER_CLASS_CONFIG, JsonSerializer.class);
-        return props;
-    }
-
-    @Bean
-    public KafkaTemplate<String, Introduce> kafkaTemplate() {
-        return new KafkaTemplate<>(producerFactory());
-    }
-}
