--- src/test/java/com/kvp/kafka/KafkaApplicationTests.java
@@ -1,13 +1,0 @@
-package com.kvp.kafka;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class KafkaApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
